refactor(antithesis): split intake HTTP modules
